Specialize DisiPriorityQueue for the 2-clauses case. (#14070)
Disjunctions with 2 clauses are rather common. Specializing this case enables
some shortcuts.
